refactor(auth): move authOptions to dedicated config file
Centralize auth configuration in a dedicated file to improve maintainability and reduce code duplication. This change also updates imports across the application to reference the new location.
